+++ src/ComptrollerStorage.sol
@@ -5,28 +5,27 @@ import "./PriceOracle.sol";
 
 contract UnitrollerAdminStorage {
     /**
-    * @notice Administrator for this contract
-    */
+     * @notice Administrator for this contract
+     */
     address public admin;
 
     /**
-    * @notice Pending administrator for this contract
-    */
+     * @notice Pending administrator for this contract
+     */
     address public pendingAdmin;
 
     /**
-    * @notice Active brains of Unitroller
-    */
+     * @notice Active brains of Unitroller
+     */
     address public comptrollerImplementation;
 
     /**
-    * @notice Pending brains of Unitroller
-    */
+     * @notice Pending brains of Unitroller
+     */
     address public pendingComptrollerImplementation;
 }
 
 contract ComptrollerVXStorage is UnitrollerAdminStorage {
-
     /**
      * @notice Oracle which gives the price of any given asset
      */
@@ -35,17 +34,17 @@ contract ComptrollerVXStorage is UnitrollerAdminStorage {
     /**
      * @notice Multiplier used to calculate the maximum repayAmount when liquidating a borrow
      */
-    uint public closeFactorMantissa;
+    uint256 public closeFactorMantissa;
 
     /**
      * @notice Multiplier representing the discount on collateral that a liquidator receives
      */
-    uint public liquidationIncentiveMantissa;
+    uint256 public liquidationIncentiveMantissa;
 
     /**
      * @notice Max number of assets a single account can participate in (borrow or use as collateral)
      */
-    uint public maxAssets;
+    uint256 public maxAssets;
 
     /**
      * @notice Per-account mapping of "assets you are in", capped by maxAssets
@@ -55,17 +54,14 @@ contract ComptrollerVXStorage is UnitrollerAdminStorage {
     struct Market {
         /// @notice Whether or not this market is listed
         bool isListed;
-
         /**
          * @notice Multiplier representing the most one can borrow against their collateral in this market.
          *  For instance, 0.9 to allow borrowing 90% of collateral value.
          *  Must be between 0 and 1, and stored as a mantissa.
          */
-        uint collateralFactorMantissa;
-
+        uint256 collateralFactorMantissa;
         /// @notice Per-market mapping of "accounts in this asset"
         mapping(address => bool) accountMembership;
-
         /// @notice Whether or not this market receives WELL
         bool isWelled;
     }
@@ -75,7 +71,6 @@ contract ComptrollerVXStorage is UnitrollerAdminStorage {
      * @dev Used e.g. to determine if a market is supported
      */
     mapping(address => Market) public markets;
-
 
     /**
      * @notice The Pause Guardian can pause certain actions as a safety mechanism.
@@ -97,39 +92,38 @@ contract ComptrollerVXStorage is UnitrollerAdminStorage {
     address public borrowCapGuardian;
 
     // @notice Borrow caps enforced by borrowAllowed for each mToken address. Defaults to zero which corresponds to unlimited borrowing.
-    mapping(address => uint) public borrowCaps;
+    mapping(address => uint256) public borrowCaps;
 
     struct RewardMarketState {
         /// @notice The market's last updated rewardBorrowIndex or rewardSupplyIndex
         uint224 index;
-
         /// @notice The block timestamp the index was last updated at
         uint32 timestamp;
     }
 
     /// @notice The rate at which the flywheel distributes reward, per timestamp
-    mapping(uint8 => uint) rewardRate;
+    mapping(uint8 => uint256) rewardRate;
 
     /// @notice The portion of supply reward rate that each market currently receives
-    mapping(uint8 => mapping(address => uint)) public supplyRewardSpeeds;
+    mapping(uint8 => mapping(address => uint256)) public supplyRewardSpeeds;
 
     /// @notice The portion of borrow reward rate that each market currently receives
-    mapping(uint8 => mapping(address => uint)) public borrowRewardSpeeds;
+    mapping(uint8 => mapping(address => uint256)) public borrowRewardSpeeds;
 
     /// @notice The WELL/GLMR market supply state for each market
     mapping(uint8 => mapping(address => RewardMarketState)) public rewardSupplyState;
 
     /// @notice The WELL/GLMR market borrow state for each market
-    mapping(uint8 =>mapping(address => RewardMarketState)) public rewardBorrowState;
+    mapping(uint8 => mapping(address => RewardMarketState)) public rewardBorrowState;
 
     /// @notice The WELL/GLMR borrow index for each market for each supplier as of the last time they accrued reward
-    mapping(uint8 => mapping(address => mapping(address => uint))) public rewardSupplierIndex;
+    mapping(uint8 => mapping(address => mapping(address => uint256))) public rewardSupplierIndex;
 
     /// @notice The WELL/GLMR borrow index for each market for each borrower as of the last time they accrued reward
-    mapping(uint8 => mapping(address => mapping(address => uint))) public rewardBorrowerIndex;
+    mapping(uint8 => mapping(address => mapping(address => uint256))) public rewardBorrowerIndex;
 
     /// @notice The WELL/GLMR accrued but not yet transferred to each user
-    mapping(uint8 => mapping(address => uint)) public rewardAccrued;
+    mapping(uint8 => mapping(address => uint256)) public rewardAccrued;
 
     /// @notice WELL token contract address
     address public wellAddress;
